removing unneccessary casts

+++ src/algorithms/imageProcessing/util/RANSACAlgorithmIterations.java
@@ -170,13 +170,13 @@ public class RANSACAlgorithmIterations {
         
         if (nTruePoints >= sampleSize) {        
             for (int i = 0; i < sampleSize; ++i) {
-                factor *= (double)(nTruePoints - i)/(double)(nPoints - i);
+                factor *= (nTruePoints - i)/(double)(nPoints - i);
             }
         } else {
             //know that the number is still smaller than (1./nPoints)^sampleSize.
             //For nTruePoints of 1, looks like (1./nPoints)*(1./(nPoints-sampleSize))^(sampleSize-1)
             for (int i = 0; i < (int)nTruePoints; ++i) {
-                factor *= (double)(nTruePoints - i)/(double)(nPoints - i);
+                factor *= (nTruePoints - i)/(double)(nPoints - i);
             }
             int pow = (sampleSize - (int)nTruePoints);
             factor *= Math.pow((1./(nPoints - sampleSize)), pow);
